backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.2.0
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
python-dotenv==1.0.0